issue #43 feat(multisig-ui): bottomsheet menu
- Rename VaultDetails -> VaultMenuScreen
