+++ lib/app.dart
@@ -14,7 +14,7 @@ import 'package:coconut_vault/screens/vault_creation/vault_creation_options_scre
 import 'package:coconut_vault/screens/vault_creation/vault_name_icon_setup_screen.dart';
 import 'package:coconut_vault/screens/vault_detail/address_list_screen.dart';
 import 'package:coconut_vault/screens/vault_detail/sync_to_wallet_screen.dart';
-import 'package:coconut_vault/screens/vault_detail/vault_detail_screen.dart';
+import 'package:coconut_vault/screens/vault_detail/vault_menu_screen.dart';
 import 'package:coconut_vault/screens/vault_detail/vault_settings.dart';
 import 'package:coconut_vault/utils/router_util.dart';
 import 'package:coconut_vault/vault_list_tab.dart';
@@ -164,7 +164,7 @@ class _PowVaultAppState extends State<PowVaultApp> {
             '/vault-name-setup': (context) => const VaultNameIconSetup(),
             '/vault-details': (context) => buildScreenWithArguments(
                   context,
-                  (args) => VaultDetails(id: args['id']),
+                  (args) => VaultMenuScreen(id: args['id']),
                 ),
             '/vault-settings': (context) => buildScreenWithArguments(
                   context,
